ch04/requirements.txt
langchain==1.0.3
langchain-google-genai
langchain-openai
langchain-community==0.4.1
python-dotenv
requests==2.32.5
bs4
duckduckgo-search==7.5.0
ddgs==9.5.5
